+++ src/openscm_runner/adapters/magicc7/_run_magicc_parallel.py
@@ -12,8 +12,8 @@ import pymagicc
 import scmdata
 
 from ...settings import config
+from ..utils._parallel_process import _parallel_process
 from ._magicc_instances import _MagiccInstances
-from ._parallel_process import _parallel_process
 
 LOGGER = logging.getLogger(__name__)